+++ src/common/index.ts
@@ -28,6 +28,7 @@ export {
     RpcExceptionFilter,
     WsExceptionFilter,
     NestInterceptor,
+    INestApplicationContext,
 } from './interfaces';
 export * from './services/logger.service';
 export * from './pipes';