+++ src/common/utils/index.ts
@@ -16,3 +16,4 @@ export * from './decorators/use-interceptors.decorator';
 export * from './decorators/http-code.decorator';
 export * from './decorators/bind.decorator';
 export * from './forward-ref.util';
+export * from './decorators/create-route-param-metadata.decorator';