+++ src/core/index.ts
@@ -11,3 +11,4 @@ export { ModuleRef } from './injector/module-ref';
 export * from './services/reflector.service';
 export * from './nest-factory';
 export * from './nest-application';
+export * from './nest-application-context';